pyproject.toml
[project]
name = "vj-gen"
version = "0.1.0"
description = "Multi-model VJ generation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "boto3>=1.34.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "demucs>=4.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.8.0",
    "moviepy>=1.0.3",
    "Pillow>=10.0.0",
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.struff]
line_length = 100
